agent/pyproject.toml
[project]
name = "interactive-rag-agent"
version = "1.0.0"
description = "Interactive RAG Agent demonstrating CopilotKit useAgent hook with Pydantic AI + AG-UI protocol"
authors = [{ name = "Interactive RAG Demo" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai-slim[ag-ui,openai]>=1.36.0",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "asyncpg>=0.30.0",
    "openai>=1.58.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
    "tiktoken>=0.8.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
